Move function to better file
