+++ src/zenml/model/utils.py
@@ -23,9 +23,11 @@ from zenml.exceptions import StepContextError
 from zenml.logger import get_logger
 from zenml.metadata.metadata_types import MetadataType
 from zenml.model.model import Model
-from zenml.models import ArtifactVersionResponse, ServiceUpdate
-from zenml.orchestrators.step_run_utils import (
-    link_artifact_version_to_model_version,
+from zenml.models import (
+    ArtifactVersionResponse,
+    ModelVersionArtifactRequest,
+    ModelVersionResponse,
+    ServiceUpdate,
 )
 from zenml.steps.step_context import get_step_context
 
@@ -69,6 +71,39 @@ def log_model_metadata(
         mv = step_context.model
 
     mv.log_metadata(metadata)
+
+
+def link_artifact_version_to_model_version(
+    artifact_version: ArtifactVersionResponse,
+    model_version: ModelVersionResponse,
+    artifact_config: Optional[ArtifactConfig] = None,
+) -> None:
+    """Link an artifact version to a pipeline version.
+
+    Args:
+        artifact_version: The artifact version to link.
+        model_version: The model version to link.
+        artifact_config: Output artifact configuration.
+    """
+    if artifact_config:
+        is_model_artifact = artifact_config.is_model_artifact
+        is_deployment_artifact = artifact_config.is_deployment_artifact
+    else:
+        is_model_artifact = False
+        is_deployment_artifact = False
+
+    client = Client()
+    client.zen_store.create_model_version_artifact_link(
+        ModelVersionArtifactRequest(
+            user=client.active_user.id,
+            workspace=client.active_workspace.id,
+            artifact_version=artifact_version.id,
+            model=model_version.model.id,
+            model_version=model_version.id,
+            is_model_artifact=is_model_artifact,
+            is_deployment_artifact=is_deployment_artifact,
+        )
+    )
 
 
 def link_artifact_to_model(
